% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edss_calculator.R
\name{edss_calculation}
\alias{edss_calculation}
\title{EDSS Calculator}
\usage{
edss_calculation(
  score_pyramidal,
  score_cerebellar,
  score_brain_stem,
  score_sensory,
  score_bowel_bladder,
  score_visual,
  score_mental,
  score_ambulation,
  imed = F
)
}
\arguments{
\item{score_pyramidal}{Pyramidal functional subsystem ranges 0 - 6}

\item{score_cerebellar}{Cerebellar functional subsystem ranges 0 - 5}

\item{score_brain_stem}{Brain Stem functional subsystem ranges 0 - 5}

\item{score_sensory}{Sensory functional subsystem ranges 0 - 6}

\item{score_bowel_bladder}{Bowel and Bladder functional subsystem ranges 0 - 6}

\item{score_visual}{Visual functional subsystem ranges 0 - 6}

\item{score_mental}{Cerebral functional subsystem ranges 0 - 5}

\item{score_ambulation}{Ambulation score that ranges 0 - 16 (for imed = F) or 0 - 13 (for imed = T)}

\item{imed}{This parameter allows to choose whether to use original Neurostatus range 0 - 16 for ambulation score (default; imed = F). By setting imed = T you can use shortened version with range 0 - 13 in ambulation score. In this cace the original ambulation subscores of 5-7 and 8-9 are merged into scores 5 and 6 respectively.}
}
\value{
EDSS total score (ranging between 0 and 10)
}
\description{
R Package for calculation of the total EDSS based on scores from functional subsystems
}
\examples{
edss_total_01 <- edss_calculation(3, 1, 2, 3, 2, 1, 1, 6)
edss_total_02 <- edss_calculation(3, 1, 2, 3, 2, 1, 1, 6, imed = T)
}
