% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edss_calculator.R
\name{edss_calculation}
\alias{edss_calculation}
\title{EDSS Calculator}
\usage{
edss_calculation(
  score_pyramidal,
  score_cerebellar,
  score_brain_stem,
  score_sensory,
  score_bowel_bladder,
  score_visual,
  score_mental,
  score_ambulation,
  neurostatus = T
)
}
\arguments{
\item{score_pyramidal}{Pyramidal functional subsystem ranges 0 - 6}

\item{score_cerebellar}{Cerebellar functional subsystem ranges 0 - 5}

\item{score_brain_stem}{Brain Stem functional subsystem ranges 0 - 5}

\item{score_sensory}{Sensory functional subsystem ranges 0 - 6}

\item{score_bowel_bladder}{Bowel and Bladder functional subsystem ranges 0 - 6}

\item{score_visual}{Visual functional subsystem ranges 0 - 6}

\item{score_mental}{Cerebral functional subsystem ranges 0 - 5}

\item{score_ambulation}{Ambulation score that ranges 0 - 16 (for neurostatus = T) or 0 - 12(13) (for neurostatus = F)}

\item{neurostatus}{This parameter allows to choose whether to use Neurostatus range 0 - 16 for ambulation score (default; neurostatus = T). By setting neurostatus = F you can use the original Kurtzke's scale with range 0 - 12(13) in ambulation score. In this cace the ambulation subscores of 5-7 and 8-9 from Neurostatus are merged into scores 5 and 6 respectively; Also ambulation score of 1 automatically leads to total EDSS = 4. This situation is also handled differently in te case of Neurostatus}
}
\value{
EDSS total score (ranging between 0 and 10)
}
\description{
R Package for calculation of the total EDSS based on scores from functional subsystems
}
\examples{
edss_total_01 <- edss_calculation(3, 1, 2, 3, 2, 1, 1, 6)
edss_total_02 <- edss_calculation(3, 1, 2, 3, 2, 1, 1, 6, neurostatus = F)
}
