% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edss_calculator.R
\name{edss_calculation}
\alias{edss_calculation}
\title{EDSS Calculator}
\usage{
edss_calculation(
  score_pyramidal,
  score_cerebellar,
  score_brain_stem,
  score_sensory,
  score_bowel_bladder,
  score_visual,
  score_mental,
  score_ambulation,
  edss_method = "kurtzke"
)
}
\arguments{
\item{score_pyramidal}{Pyramidal functional subsystem ranges 0 - 6}

\item{score_cerebellar}{Cerebellar functional subsystem ranges 0 - 5}

\item{score_brain_stem}{Brain Stem functional subsystem ranges 0 - 5}

\item{score_sensory}{Sensory functional subsystem ranges 0 - 6}

\item{score_bowel_bladder}{Bowel and Bladder functional subsystem ranges 0 - 6}

\item{score_visual}{Visual functional subsystem ranges 0 - 6}

\item{score_mental}{Cerebral functional subsystem ranges 0 - 5}

\item{score_ambulation}{Ambulation score that ranges 0 - 16 (for edss_method = "neurostatus_16") or 0 - 12(13) (for edss_method = "neurostatus_12" or "kurtzke"))}

\item{edss_method}{This parameter allows to choose the method of EDSS calculation. There are three possible parameters: 1) "kurtzke" (default): The original procedure by Kurtzke; 2) "neurostatus_12" Neurostatus evaluation, in this case with 12(13) grades for ambulation scale; 3) "neurostatus_16" Same Neurostatus evaluation but in this case with 16 grades for ambulation scale.}
}
\value{
EDSS total score (ranging between 0 and 10)
}
\description{
R Package for calculation of the total EDSS based on scores from functional subsystems
}
\examples{
edss_total_01 <- edss_calculation(3, 1, 2, 3, 2, 1, 1, 6)
edss_total_02 <- edss_calculation(3, 1, 2, 3, 2, 1, 1, 6, edss_method = "neurostatus_12")
edss_total_03 <- edss_calculation(3, 1, 2, 3, 2, 1, 1, 6, edss_method = "neurostatus_16")
edss_total_04 <- edss_calculation(3, 1, 2, 3, 2, 1, 1, 6, edss_method = "kurtzke")
}
